DBZ-7083 Introduce strategy for MySQL and MariaDB
