+++ debezium-connector-mysql/pom.xml
@@ -176,6 +176,7 @@
         <mysql.init.timeout>60000</mysql.init.timeout> <!-- 60 seconds -->
         <apicurio.port>8080</apicurio.port>
         <apicurio.init.timeout>60000</apicurio.init.timeout> <!-- 60 seconds -->
+        <mysql.database.adapter>mysql</mysql.database.adapter>
         <mysql.database.protocol>jdbc:mysql</mysql.database.protocol>
         <mysql.database.jdbc.driver>com.mysql.cj.jdbc.Driver</mysql.database.jdbc.driver>
         <!--
@@ -548,6 +549,7 @@
                         <database.replica.hostname>${docker.host.address}</database.replica.hostname>
                         <database.replica.port>${mysql.replica.port}</database.replica.port>
                         <!-- Specifies which driver to use for the tests -->
+                        <connector.adapter>${mysql.database.adapter}</connector.adapter>
                         <database.protocol>${mysql.database.protocol}</database.protocol>
                         <database.jdbc.driver>${mysql.database.jdbc.driver}</database.jdbc.driver>
                         <skipLongRunningTests>${skipLongRunningTests}</skipLongRunningTests>
@@ -696,6 +698,7 @@
                                 <database.replica.port>${mysql.port}</database.replica.port>
                                 <database.ssl.mode>disabled</database.ssl.mode>
                                 <!-- Specifies which driver to use for the tests -->
+                                <connector.adapter>${mysql.database.adapter}</connector.adapter>
                                 <database.protocol>${mysql.database.protocol}</database.protocol>
                                 <database.jdbc.driver>${mysql.database.jdbc.driver}</database.jdbc.driver>
                                 <skipLongRunningTests>false</skipLongRunningTests>
@@ -716,6 +719,7 @@
                                         <database.port>${mysql.port}</database.port>
                                         <database.replica.port>${mysql.port}</database.replica.port>
                                         <!-- Specifies which driver to use for the tests -->
+                                        <connector.adapter>${mysql.database.adapter}</connector.adapter>
                                         <database.protocol>${mysql.database.protocol}</database.protocol>
                                         <database.jdbc.driver>${mysql.database.jdbc.driver}</database.jdbc.driver>
                                     </systemPropertyVariables>
@@ -732,6 +736,7 @@
                                         <database.port>${mysql.gtid.port}</database.port>
                                         <database.replica.port>${mysql.gtid.replica.port}</database.replica.port>
                                         <!-- Specifies which driver to use for the tests -->
+                                        <connector.adapter>${mysql.database.adapter}</connector.adapter>
                                         <database.protocol>${mysql.database.protocol}</database.protocol>
                                         <database.jdbc.driver>${mysql.database.jdbc.driver}</database.jdbc.driver>
                                     </systemPropertyVariables>
@@ -768,6 +773,7 @@
                                         <database.port>${mysql.ssl.port}</database.port>
                                         <database.replica.port>${mysql.ssl.port}</database.replica.port>
                                         <!-- Specifies which driver to use for the tests -->
+                                        <connector.adapter>${mysql.database.adapter}</connector.adapter>
                                         <database.protocol>${mysql.database.protocol}</database.protocol>
                                         <database.jdbc.driver>${mysql.database.jdbc.driver}</database.jdbc.driver>
                                     </systemPropertyVariables>
@@ -932,23 +938,9 @@
                 <activeByDefault>false</activeByDefault>
             </activation>
             <properties>
-                <docker.dbs>debezium/mysql-server-test-database</docker.dbs>
                 <mysql.server.image.source>${mariadb.server.image.source}</mysql.server.image.source>
                 <version.mysql.server>${version.mariadb.server}</version.mysql.server>
-                <mysql.database.protocol>jdbc:mariadb</mysql.database.protocol>
-                <mysql.database.jdbc.driver>org.mariadb.jdbc.Driver</mysql.database.jdbc.driver>
-            </properties>
-        </profile>
-        <!-- Profile for MariaDB SQL support -->
-        <profile>
-            <id>mariadb-ssl</id>
-            <activation>
-                <activeByDefault>false</activeByDefault>
-            </activation>
-            <properties>
-                <docker.dbs>debezium/mysql-server-test-database-ssl</docker.dbs>
-                <mysql.server.image.source>${mariadb.server.image.source}</mysql.server.image.source>
-                <version.mysql.server>${version.mariadb.server}</version.mysql.server>
+                <mysql.database.adapter>mariadb</mysql.database.adapter>
                 <mysql.database.protocol>jdbc:mariadb</mysql.database.protocol>
                 <mysql.database.jdbc.driver>org.mariadb.jdbc.Driver</mysql.database.jdbc.driver>
             </properties>
@@ -1016,6 +1008,7 @@
                                 <!-- Specifies which driver to use for the tests -->
                                 <database.protocol>${mysql.database.protocol}</database.protocol>
                                 <database.jdbc.driver>${mysql.database.jdbc.driver}</database.jdbc.driver>
+                                <connector.adapter>${mysql.database.adapter}</connector.adapter>
                                 <skipLongRunningTests>false</skipLongRunningTests>
                                 <isAssemblyProfileActive>true</isAssemblyProfileActive>
                             </systemPropertyVariables>
@@ -1096,6 +1089,7 @@
                                 <!-- Specifies which driver to use for the tests -->
                                 <database.protocol>${mysql.database.protocol}</database.protocol>
                                 <database.jdbc.driver>${mysql.database.jdbc.driver}</database.jdbc.driver>
+                                <connector.adapter>${mysql.database.adapter}</connector.adapter>
                                 <skipLongRunningTests>false</skipLongRunningTests>
                                 <isAssemblyProfileActive>true</isAssemblyProfileActive>
                             </systemPropertyVariables>
@@ -1174,6 +1168,7 @@
                                 <!-- Specifies which driver to use for the tests -->
                                 <database.protocol>${mysql.database.protocol}</database.protocol>
                                 <database.jdbc.driver>${mysql.database.jdbc.driver}</database.jdbc.driver>
+                                <connector.adapter>${mysql.database.adapter}</connector.adapter>
                                 <skipLongRunningTests>false</skipLongRunningTests>
                                 <isAssemblyProfileActive>true</isAssemblyProfileActive>
                             </systemPropertyVariables>
@@ -1252,6 +1247,7 @@
                                 <!-- Specifies which driver to use for the tests -->
                                 <database.protocol>${mysql.database.protocol}</database.protocol>
                                 <database.jdbc.driver>${mysql.database.jdbc.driver}</database.jdbc.driver>
+                                <connector.adapter>${mysql.database.adapter}</connector.adapter>
                                 <skipLongRunningTests>false</skipLongRunningTests>
                                 <isAssemblyProfileActive>true</isAssemblyProfileActive>
                                 <database.ssl.truststore>${project.basedir}/src/test/resources/ssl/truststore</database.ssl.truststore>
