+++ debezium-connector-mysql/src/test/java/io/debezium/connector/mysql/GtidSetTest.java
@@ -16,8 +16,9 @@ import java.util.stream.Collectors;
 
 import org.junit.Test;
 
-import io.debezium.connector.mysql.GtidSet.Interval;
-import io.debezium.connector.mysql.GtidSet.UUIDSet;
+import io.debezium.connector.mysql.strategy.mysql.MySqlGtidSet;
+import io.debezium.connector.mysql.strategy.mysql.MySqlGtidSet.Interval;
+import io.debezium.connector.mysql.strategy.mysql.MySqlGtidSet.UUIDSet;
 import io.debezium.util.Collect;
 
 /**
@@ -28,11 +29,11 @@ public class GtidSetTest {
 
     private static final String UUID1 = "24bc7850-2c16-11e6-a073-0242ac110002";
 
-    private GtidSet gtids;
+    private MySqlGtidSet gtids;
 
     @Test
     public void shouldCreateSetWithSingleInterval() {
-        gtids = new GtidSet(UUID1 + ":1-191");
+        gtids = new MySqlGtidSet(UUID1 + ":1-191");
         asertIntervalCount(UUID1, 1);
         asertIntervalExists(UUID1, 1, 191);
         asertFirstInterval(UUID1, 1, 191);
@@ -42,7 +43,7 @@ public class GtidSetTest {
 
     @Test
     public void shouldCollapseAdjacentIntervals() {
-        gtids = new GtidSet(UUID1 + ":1-191:192-199");
+        gtids = new MySqlGtidSet(UUID1 + ":1-191:192-199");
         asertIntervalCount(UUID1, 1);
         asertIntervalExists(UUID1, 1, 199);
         asertFirstInterval(UUID1, 1, 199);
@@ -52,7 +53,7 @@ public class GtidSetTest {
 
     @Test
     public void shouldNotCollapseNonAdjacentIntervals() {
-        gtids = new GtidSet(UUID1 + ":1-191:193-199");
+        gtids = new MySqlGtidSet(UUID1 + ":1-191:193-199");
         asertIntervalCount(UUID1, 2);
         asertFirstInterval(UUID1, 1, 191);
         asertLastInterval(UUID1, 193, 199);
@@ -61,7 +62,7 @@ public class GtidSetTest {
 
     @Test
     public void shouldCreateWithMultipleIntervals() {
-        gtids = new GtidSet(UUID1 + ":1-191:193-199:1000-1033");
+        gtids = new MySqlGtidSet(UUID1 + ":1-191:193-199:1000-1033");
         asertIntervalCount(UUID1, 3);
         asertFirstInterval(UUID1, 1, 191);
         asertIntervalExists(UUID1, 193, 199);
@@ -71,7 +72,7 @@ public class GtidSetTest {
 
     @Test
     public void shouldCreateWithMultipleIntervalsThatMayBeAdjacent() {
-        gtids = new GtidSet(UUID1 + ":1-191:192-199:1000-1033:1035-1036:1038-1039");
+        gtids = new MySqlGtidSet(UUID1 + ":1-191:192-199:1000-1033:1035-1036:1038-1039");
         asertIntervalCount(UUID1, 4);
         asertFirstInterval(UUID1, 1, 199);
         asertIntervalExists(UUID1, 1000, 1033);
@@ -82,19 +83,19 @@ public class GtidSetTest {
 
     @Test
     public void shouldCorrectlyDetermineIfSimpleGtidSetIsContainedWithinAnother() {
-        gtids = new GtidSet("7c1de3f2-3fd2-11e6-9cdc-42010af000bc:1-41");
-        assertThat(gtids.isContainedWithin(new GtidSet("7c1de3f2-3fd2-11e6-9cdc-42010af000bc:1-41"))).isTrue();
-        assertThat(gtids.isContainedWithin(new GtidSet("7c1de3f2-3fd2-11e6-9cdc-42010af000bc:1-42"))).isTrue();
-        assertThat(gtids.isContainedWithin(new GtidSet("7c1de3f2-3fd2-11e6-9cdc-42010af000bc:2-41"))).isFalse();
-        assertThat(gtids.isContainedWithin(new GtidSet("7145bf69-d1ca-11e5-a588-0242ac110004:1"))).isFalse();
+        gtids = new MySqlGtidSet("7c1de3f2-3fd2-11e6-9cdc-42010af000bc:1-41");
+        assertThat(gtids.isContainedWithin(new MySqlGtidSet("7c1de3f2-3fd2-11e6-9cdc-42010af000bc:1-41"))).isTrue();
+        assertThat(gtids.isContainedWithin(new MySqlGtidSet("7c1de3f2-3fd2-11e6-9cdc-42010af000bc:1-42"))).isTrue();
+        assertThat(gtids.isContainedWithin(new MySqlGtidSet("7c1de3f2-3fd2-11e6-9cdc-42010af000bc:2-41"))).isFalse();
+        assertThat(gtids.isContainedWithin(new MySqlGtidSet("7145bf69-d1ca-11e5-a588-0242ac110004:1"))).isFalse();
     }
 
     @Test
     public void shouldCorrectlyDetermineIfComplexGtidSetIsContainedWithinAnother() {
-        GtidSet connector = new GtidSet("036d85a9-64e5-11e6-9b48-42010af0000c:1-2,"
+        MySqlGtidSet connector = new MySqlGtidSet("036d85a9-64e5-11e6-9b48-42010af0000c:1-2,"
                 + "7145bf69-d1ca-11e5-a588-0242ac110004:1-3200,"
                 + "7c1de3f2-3fd2-11e6-9cdc-42010af000bc:1-41");
-        GtidSet server = new GtidSet("036d85a9-64e5-11e6-9b48-42010af0000c:1-2,"
+        MySqlGtidSet server = new MySqlGtidSet("036d85a9-64e5-11e6-9b48-42010af0000c:1-2,"
                 + "7145bf69-d1ca-11e5-a588-0242ac110004:1-3202,"
                 + "7c1de3f2-3fd2-11e6-9cdc-42010af000bc:1-41");
         assertThat(connector.isContainedWithin(server)).isTrue();
@@ -102,12 +103,12 @@ public class GtidSetTest {
 
     @Test
     public void shouldCorrectlyDetermineIfComplexGtidSetWithVariousLineSeparatorsIsContainedWithinAnother() {
-        GtidSet connector = new GtidSet("036d85a9-64e5-11e6-9b48-42010af0000c:1-2,"
+        GtidSet connector = new MySqlGtidSet("036d85a9-64e5-11e6-9b48-42010af0000c:1-2,"
                 + "7145bf69-d1ca-11e5-a588-0242ac110004:1-3200,"
                 + "7c1de3f2-3fd2-11e6-9cdc-42010af000bc:1-41");
         Arrays.stream(new String[]{ "\r\n", "\n", "\r" })
                 .forEach(separator -> {
-                    GtidSet server = new GtidSet("036d85a9-64e5-11e6-9b48-42010af0000c:1-2," + separator +
+                    GtidSet server = new MySqlGtidSet("036d85a9-64e5-11e6-9b48-42010af0000c:1-2," + separator +
                             "7145bf69-d1ca-11e5-a588-0242ac110004:1-3202," + separator +
                             "7c1de3f2-3fd2-11e6-9cdc-42010af000bc:1-41");
                     assertThat(connector.isContainedWithin(server)).isTrue();
@@ -122,12 +123,12 @@ public class GtidSetTest {
         Collection<String> keepers = Collect.arrayListOf("036d85a9-64e5-11e6-9b48-42010af0000c",
                 "7c1de3f2-3fd2-11e6-9cdc-42010af000bc",
                 "wont-be-found");
-        GtidSet original = new GtidSet(gtidStr);
+        MySqlGtidSet original = new MySqlGtidSet(gtidStr);
         assertThat(original.forServerWithId("036d85a9-64e5-11e6-9b48-42010af0000c")).isNotNull();
         assertThat(original.forServerWithId("7c1de3f2-3fd2-11e6-9cdc-42010af000bc")).isNotNull();
         assertThat(original.forServerWithId("7145bf69-d1ca-11e5-a588-0242ac110004")).isNotNull();
 
-        GtidSet filtered = original.retainAll(keepers::contains);
+        MySqlGtidSet filtered = original.retainAll(keepers::contains);
         List<String> actualUuids = filtered.getUUIDSets().stream().map(UUIDSet::getUUID).collect(Collectors.toList());
         assertThat(keepers.containsAll(actualUuids)).isTrue();
         assertThat(filtered.forServerWithId("7145bf69-d1ca-11e5-a588-0242ac110004")).isNull();
@@ -144,11 +145,11 @@ public class GtidSetTest {
         String diff = "036d85a9-64e5-11e6-9b48-42010af0000c:21,"
                 + "7145bf69-d1ca-11e5-a588-0242ac110004:4500,"
                 + "7c1de3f2-3fd2-11e6-9cdc-42010af000bc:1-4:9-11:19-24:66-70:80-100";
-        GtidSet gtidSet1 = new GtidSet(gtidStr1);
-        GtidSet gtidSet2 = new GtidSet(gtidStr2);
+        MySqlGtidSet gtidSet1 = new MySqlGtidSet(gtidStr1);
+        MySqlGtidSet gtidSet2 = new MySqlGtidSet(gtidStr2);
 
-        GtidSet gtidSetDiff = gtidSet2.subtract(gtidSet1);
-        GtidSet expectedDiff = new GtidSet(diff);
+        MySqlGtidSet gtidSetDiff = gtidSet2.subtract(gtidSet1);
+        MySqlGtidSet expectedDiff = new MySqlGtidSet(diff);
         assertThat(gtidSetDiff).isEqualTo(expectedDiff);
     }
