+++ debezium-connector-mysql/src/test/java/io/debezium/connector/mysql/MySqlDatabaseSchemaTest.java
@@ -68,7 +68,8 @@ public class MySqlDatabaseSchemaTest {
                 BigIntUnsignedMode.LONG,
                 BinaryHandlingMode.BYTES,
                 MySqlValueConverters::adjustTemporal,
-                MySqlValueConverters::defaultParsingErrorHandler);
+                MySqlValueConverters::defaultParsingErrorHandler,
+                connectorConfig.getConnectorAdapter());
         return new MySqlDatabaseSchema(
                 connectorConfig,
                 mySqlValueConverters,
